Rename a file.
