+++ cpp/dolfinx/mesh/CMakeLists.txt
@@ -8,7 +8,7 @@ set(HEADERS_mesh
   ${CMAKE_CURRENT_SOURCE_DIR}/Geometry.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshIterator.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/PermutationInfo.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/PermutationComputation.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshQuality.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshValueCollection.h
@@ -23,7 +23,7 @@ target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/Mesh.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Geometry.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/PermutationInfo.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/PermutationComputation.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshQuality.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/TopologyComputation.cpp
