+++ cpp/dolfinx/mesh/PermutationComputation.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier:    LGPL-3.0-or-later
 
-#include "PermutationInfo.h"
+#include "PermutationComputation.h"
 #include <bitset>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <dolfinx/mesh/Topology.h>
@@ -243,7 +243,8 @@ compute_face_permutations(const mesh::Topology& topology)
 //-----------------------------------------------------------------------------
 std::pair<Eigen::Array<std::uint8_t, Eigen::Dynamic, Eigen::Dynamic>,
           Eigen::Array<std::uint32_t, Eigen::Dynamic, 1>>
-PermutationInfo::compute_entity_permutations(const mesh::Topology& topology)
+PermutationComputation::compute_entity_permutations(
+    const mesh::Topology& topology)
 {
   const int tdim = topology.dim();
   const CellType cell_type = topology.cell_type();
